Refactor documentation command
